+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/src/encryption/TEA.cpp
@@ -91,11 +91,11 @@ void CTEA::Signature(char* pcSig)
 	//3+16+1+1+1
 	char acSigData[23] = {0};
 	strcat(acSigData, "TEA");
-	int iLen = strlen(acSigData);
+	int iLen = static_cast<int>(strlen(acSigData));
 	memcpy(acSigData+iLen, m_apKey.get(), m_keylength);
 	sprintf(acSigData+iLen+m_keylength, "%d%d", m_iMode, m_iPadding);
 	CSHA oSHA;
-	oSHA.AddData(acSigData, strlen(acSigData));
+	oSHA.AddData(acSigData, static_cast<int>(strlen(acSigData)));
 	oSHA.FinalDigest(pcSig);
 }
